Views folder deleted, class UserViewPage moved to Users folder
